Split the "Deploy a Cluster" section
The "Deploy a Cluster" section includes subsections related to choosing
a Teleport edition as well as subsections related to deploying a
production cluster. The problem with this approach is that it leaves no
room for adding introductory guides to help users with these two
purposes.

By splitting the "Deploy a Cluster" section into top-level sections, one
related to choosing an edition and the other to dpeloying to production,
we can add introductory pages that put the subsections of these sections
in context.