Create "Configuring your Cluster" docs subsection
Closes #25418

Help make infrastructure-as-code approaches more prominent in the
Teleport docs by separating the Operator and Terraform guides into a
"Configuring your Cluster" subsection of the "Management" section. This
also adds an introductory page that explains the three major tools we
expose for interacting with a Teleport cluster (not counting the API).

This subsection also gives us a place to put other guides to using
Terraform and the Operator, e.g., if we add guides to using these tools
with popular GitOps platforms.